backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "province-legal-os-backend"
version = "0.1.0"
description = "AI-native legal OS backend with FastAPI and AWS integration"
authors = [
    {name = "Province Team", email = "team@province.com"},
]
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "boto3>=1.34.0",
    "botocore>=1.34.0",
    "python-multipart>=0.0.6",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "websockets>=12.0",
    "httpx>=0.25.0",
    "structlog>=23.2.0",
    "tenacity>=8.2.0",
    "pyyaml>=6.0.1",
    "jinja2>=3.1.2",
    "python-dateutil>=2.8.2",
    "croniter>=2.0.1",
]
requires-python = ">=3.11"

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.7.0",
    "pre-commit>=3.5.0",
    "httpx>=0.25.0",
    "moto[all]>=4.2.0",
]
cdk = [
    "aws-cdk-lib>=2.110.0",
    "constructs>=10.3.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true